requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

langchain
langgraph
langchain-core
langchain_text_splitters

openai
tiktoken

faiss-cpu
sentence-transformers

sqlalchemy
pymysql

streamlit

pytest
pypdf